Change cert.get_fingerprint to str type
A recent commit in M2crypto changed the return type of X509.get_fingerprint() from 'str' to 'unicode' in Python2.
https://gitlab.com/m2crypto/m2crypto/commit/b977909fc93caa599e52943be25b7f6042a4c70b

However the 2nd parameter in X509.new_extension expect a 'str' type instead of 'unicode' type. This commit convert the return value of cert.get_fingerprint to a 'str' so it can continue to work.